Move detection to map module, add command
