+++ src/main/java/com/wynntils/modules/map/events/ClientEvents.java
@@ -7,10 +7,12 @@ package com.wynntils.modules.map.events;
 import com.wynntils.Reference;
 import com.wynntils.core.events.custom.GameEvent;
 import com.wynntils.core.events.custom.GuiOverlapEvent;
+import com.wynntils.core.events.custom.LocationEvent;
 import com.wynntils.core.events.custom.PacketEvent;
 import com.wynntils.core.framework.interfaces.Listener;
 import com.wynntils.core.framework.rendering.colors.CommonColors;
 import com.wynntils.core.utils.objects.Location;
+import com.wynntils.modules.map.instances.LabelBake;
 import com.wynntils.modules.core.managers.CompassManager;
 import com.wynntils.modules.map.MapModule;
 import com.wynntils.modules.map.configs.MapConfig;
@@ -26,15 +28,29 @@ import net.minecraft.block.state.IBlockState;
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.entity.EntityPlayerSP;
 import net.minecraft.entity.Entity;
+import net.minecraft.entity.passive.EntityVillager;
 import net.minecraft.network.play.server.SPacketAdvancementInfo;
 import net.minecraft.util.math.BlockPos;
+import net.minecraft.util.text.TextFormatting;
 import net.minecraftforge.client.event.RenderWorldLastEvent;
 import net.minecraftforge.event.entity.player.PlayerInteractEvent;
 import net.minecraftforge.fml.common.eventhandler.EventPriority;
 import net.minecraftforge.fml.common.eventhandler.SubscribeEvent;
 import net.minecraftforge.fml.common.gameevent.TickEvent;
 
+import java.util.regex.Matcher;
+import java.util.regex.Pattern;
+
+import static com.wynntils.modules.core.events.ClientEvents.*;
+
 public class ClientEvents implements Listener {
+    private static final Pattern MOB_LABEL = Pattern.compile("^.*\\[Lv. [0-9]+\\]$");
+    private static final Pattern HEALTH_LABEL = Pattern.compile("^\\[\\|+[0-9]+\\|+\\]$");
+    private static final Pattern TOTEM_LABEL = Pattern.compile("^[0-9]+s|\\+[0-9]+❤/s$");
+    private static final Pattern GATHERING_LABEL = Pattern.compile("^. [ⒸⒷⒿⓀ] .* Lv. Min: [0-9]+$");
+    private static final Pattern RESOURCE_LABEL = Pattern.compile("^(?:Right|Left)-Click for .*$");
+    private static final Pattern WYBEL_OWNER = Pattern.compile("^\\[.*\\]$");
+    private static final Pattern WYBEL_LEVEL = Pattern.compile("^Lv. [0-9]+.*$");
 
     BlockPos lastLocation = null;
 
@@ -123,6 +139,62 @@ public class ClientEvents implements Listener {
     public void receiveAdvancements(PacketEvent.Incoming<SPacketAdvancementInfo> event) {
         // can be done async without problems
         GuildResourceManager.processAdvancements(event.getPacket());
+    }
+
+
+    @SubscribeEvent
+    public void onTestLabelFound(LocationEvent.LabelFoundEvent event) {
+        String label = TextFormatting.getTextWithoutFormattingCodes(event.getLabel());
+        Location location = event.getLocation();
+
+        Matcher m = MOB_LABEL.matcher(label);
+        if (m.find()) return;
+
+        Matcher m2 = HEALTH_LABEL.matcher(label);
+        if (m2.find()) return;
+
+        Matcher m3 = MOB_DAMAGE.matcher(label);
+        if (m3.find()) return;
+
+        Matcher m4 = GATHERING_STATUS.matcher(label);
+        if (m4.find()) return;
+
+        Matcher m5 = GATHERING_RESOURCE.matcher(label);
+        if (m5.find()) return;
+
+        Matcher m6 = TOTEM_LABEL.matcher(label);
+        if (m6.find()) return;
+
+        Matcher m7 = GATHERING_LABEL.matcher(label);
+        if (m7.find()) return;
+
+        Matcher m8 = RESOURCE_LABEL.matcher(label);
+        if (m8.find()) return;
+
+        Matcher m9 = WYBEL_OWNER.matcher(label);
+        if (m9.find()) return;
+
+        Matcher m10 = WYBEL_LEVEL.matcher(label);
+        if (m10.find()) return;
+
+        LabelBake.handleLabel(label, location);
+    }
+
+    @SubscribeEvent
+    public void onTestEntityLabelFound(LocationEvent.EntityLabelFoundEvent event) {
+        String name = TextFormatting.getTextWithoutFormattingCodes(event.getLabel());
+        Location location = event.getLocation();
+        Entity entity = event.getEntity();
+
+        Matcher m = MOB_LABEL.matcher(name);
+        if (m.find()) return;
+
+        Matcher m2 = HEALTH_LABEL.matcher(name);
+        if (m2.find()) return;
+
+        if (!(entity instanceof EntityVillager)) return;
+
+        LabelBake.handleNpc(name, location);
     }
 
 }
